Position can be updated from the timeline's cursor, but some display bug are still presents...

+++ src/gui/TracksRuler.cpp
@@ -43,7 +43,7 @@ TracksRuler::TracksRuler( TracksView* tracksView, QWidget* parent )
     setPixelPerMark( 5 );
 
     // Redraw the ruler when the cursor position change
-    connect( tracksView->tracksCursor(), SIGNAL( cursorPositionChanged(int) ),
+    connect( tracksView->tracksCursor(), SIGNAL( cursorPositionChanged(qint64) ),
              this, SLOT( update() ) );
 }
